requirements.txt
alembic==1.8.1
amqp==5.1.1
anyio==3.6.1
attrs==22.1.0
bcrypt==4.0.1
billiard==3.6.4.0
cachetools==5.2.0
celery==5.2.7
cffi==1.15.1
chardet==5.0.0
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==38.0.1
cssselect==1.2.0
cssutils==2.6.0
dnspython==2.2.1
ecdsa==0.18.0
email-validator==1.3.0
emails==0.5.15
fastapi==0.89.1
greenlet==1.1.3.post0
h11==0.14.0
idna==3.4
iniconfig==1.1.1
kombu==5.2.4
lxml==4.9.1
Mako==1.2.3
MarkupSafe==2.1.1
numpy==1.23.4
ordereddict==1.1
packaging==21.3
pandas==1.5.0
passlib==1.7.4
pluggy==1.0.0
premailer==3.10.0
prompt-toolkit==3.0.32
psycopg2-binary==2.9.5
psycopg2-pool==1.1
py==1.11.0
py-functions==1.0
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.2
pyparsing==3.0.9
pytest==7.1.3
python==3.10
python-dateutil==2.8.2
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.5
pytz==2022.4
requests==2.28.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
starlette==0.22.0
tomli==2.0.1
typing==3.7.4.3
typing_extensions==4.4.0
urllib3==1.26.12
uvicorn==0.20.0
vine==5.0.0
wcwidth==0.2.5
